program: remove vestigial conditional `absl` import
Summary:
TensorBoard used to have an optional dependency on `absl-py`, so #1240
and #1409 were careful to only import `absl` if it was installed. But as
of #1654 TensorBoard always requires `absl-py`, so we can promote the
conditional import to a normal import.

Test Plan:
Running TensorBoard with no arguments still fails as expected; running
with `--logdir` or `--version` still do the right things, too.

wchargin-branch: absl-always