lib: message.ts source of truth is ts_library (#2757)
Due to our build quirks, we compile TypeScript files into JavaScript and
commit the transpiled code in the source tree. This process recently was
modified only to apply to files that show up in the `srcs` of
`tf_web_library`. As a result, because source-of-truth of message.ts was
part of `srcs` of a tf_web_library, it broke our genrule on the
plugin_lib side. This change moves the message.ts to plugin_lib so the 
source-of-truth is the tf_ts_library.