Adjust positioning of perf plots
For multi-class models, let's put the curves for each class in a
separated line, so that it becomes easier to follow.

For both binary and multi-class models, we adjust the margins, position
of axis labels and centralize the plots.

I also renamed and cleaned up some CSS classes on the way.