sync: unblock sync by copying message.ts (#2761)
Due to the difference between internal/OSS behavior of tf_web_library,
we now copy the message.ts over to plugin_util. The issues were following:
- Background:
  - Internal tf_web_library disallows .ts in the sources. As a result, we
    compile the TypeScript source using TSC and sync in JavaScript variant.
  - We ignore this process for tf_ts_library.
- We sync in TypeScript file in plugin_lib
- plugin_util uses message.ts by copying the file over using `genrule`
- Copied TypeScript appears at the srcs of the tf_web_library.

We can make plugin_util source of truth, but in that case, tf_ts_library
would be depend on a compiled JavaScript which is another invariant
violation.

Previous attempts to alleviate the sync issue has been unsuccessful
because I was not careful. This time, I manually ran tomservo process,
then did the sync, cl/274008245.