core: allow frame_ancestors configuration from CLI
Now, with --frame_ancestors, one can change the frame-ancestors
Content-Security-Policy to one of four configurations. The
configurations are:

- none: does not allow TensorBoard to be iframed
- colab: CSP for colab only
- notebook: CSP for notebook environments like Jupyter
- unsafe: allow any sites from embedding TensorBoard

The default for now is `unsafe` since that is current behavior as of
TB 2.0.

Do note that notebook one is the same as unsafe for now because in
context of, at least, Jupyter, localhost:[port] document loads
TensorBoard in an iframe that is running at localhost:6006. Because for
CSP keyword like "self" has to match both host-uri and the port, it does
not work for us. Moreover, it is, depending on Jupyter configuration,
naturally to run one on "ip" which would make the address not
guess-able. Until we have a better solution, we will have the "unsafe"
behavior for the notebook configuration.