plugin: load plugin bootstrap from backend
background: to load the plugin module, we created an iframe whose origin
was ambiguous (to us, TB developer). It turns out, the origin of the
src-less iframe are "about:blank" even with the `<base>` and correct
`document.domain` and, as a result, service worker did not work
correctly inside. We now create the iframe content from our backend and
serve it from our origin.

changes:
1. new `/plugin.html` endpoint that requires a plugin name that matches
  ones registered on the backend.
2. no more 'strict-dynamic' as we no longer create unknown script! We
  need to trust scripts from `self` though.