ng: fix the dev CSP issue (#2782)
TensorBoard now enforces very strict CSP and this did have a negative
impact on our tf-ng-tensorboard dev mode. Previously, TensorBoard was
able to skip the Vulcanization step of tf_ng_tensorboard because we were
using rollup bundle anyways. This skipping did improve the incremental
build speed of the Angular app. However, now that CSP infrastructure
is built on top of the Vulcanization tooling (or rather, requires shasum
of all script tags to be computed ahead of time), we are now Vulcanizing
AFTER rollup.

We have longer term plans to improve our CSP/assets story which should
improve this.