colab: use proxyPort for dynamic plugin (#2798)
proxyPort is a relatively new API that lets colab proxy a server from a
port to url given by the method. It installs a ServiceWorker when
invoked so a request is proxied to the correct kernel.

This API an iframe to be nestable and, hence, enable dynamic plugin

With this change, we now removed the colab not supported page.