scalar: fix the timing issues
Repro case
----------
1. load scalars
2. simulate slower network connection (on your favorite browser or
proxy)
3. toggle on a run
4. toggle on another run
5. (depending on the dataset) you should see nothing and line chart
defaults to 0-1 scale.
6. if you double click on the line-chart, the line is shown!

Explanation
-----------
Setting: `dataToLoad` was [a, b] and "changes" to [a, b].
Before:
  1. requests for [a, b] goes out
  2. before the requests resolve, second `loadDataImpl` gets invoked and
  cancels Promises for (1).
  3. requests from (1) resolve but because Promise was cancelled, it
  does not invoke the `onLoadFinish` and domain is not reset
After:
We now invoke `onLoadFinish` if we ever invoke `loadDataCallback` even
once.
  1. requests for [a, b] goes out
  2. before the requests resolve, second `loadDataImpl` gets invoked but
  it fetches no data and Promise.all resolves on next tick. Because no
  data is fetched, it does not call the `onLoadFinish`.
  3. requests from (1) resolve and it invokes `onLoadFinish`.

Remaining "issue"
-----------------
Setting: `dataToLoad` was [a, b] and "changes" to [a, c]. Request to
fetch `c` ends before ones for `a` and `b`.
Problem: we set the domain based on `c` only but the fix will cause
major regression with regarding auto update where any data fetches will
cause the chart to rescale even if user overrode the domains.