Bump `grpcio` version to 1.24.3
Summary:
We depend on `grpc.local_server_credentials`, which was added in 1.24.0.
Installing the Pip package into a fresh virtualenv generally picks this
up, because the `grpcio >= 1.6.3` will install the latest version, but
that won’t suffice for virtualenvs that already have an old version
installed. We now explicitly depend on `1.24.3`.

Travis wasn’t seeing this problem because the `grpcio-testing` dep pulls
in a matching version of `grpcio`.

Test Plan:
None.

wchargin-branch: grpcio-1.24.3