Bump `grpcio` version to 1.24.3 (#2829)
Summary:
We depend on `grpc.local_server_credentials`, which was added in 1.24.0.
Installing the Pip package into a fresh virtualenv generally picks this
up, because the `grpcio >= 1.6.3` will install the latest version, but
that won’t suffice for virtualenvs that already have an old version
installed. We now explicitly depend on `1.24.3`.

Travis wasn’t seeing this problem because the `grpcio-testing` dep pulls
in a matching version of `grpcio`.

Test Plan:
Create a new virtualenv and install `grpcio==1.16.0` (e.g.) into it.
Then, install the TensorBoard Pip package. Run

```
python -c 'import grpc; print(grpc.local_server_credentials)'
```

and note that this now passes instead of failing with `AttributeError`.

wchargin-branch: grpcio-1.24.3