data load: prevent JSE when canceling async
Polymer.prototype.async without a wait time defaults to next microtask.
Polymer implemented custom microtask scheduling which is not impervious
to some edge case.

Polymer has some internal counter for the microtasks and the `async`
call returns (contrived for illustration) `currentMicroTaskCounter + i`.
When canelling the async, it checks whether the handle is in between
`currentMicroTaskCounter` and `currentMicroTaskCounter` + [numAsyncs]
and it expects callback to exist when it is in between. The problem
happened when a handle is cancelled twice (our `reset` method gets
called more than once in a microtask for hparams) as it cannot find a
callback for the handle.