ng: load material icon sources (#2830)
mat-icon allows many options for loading the assets. For TensorBoard, we
would like to use, instead of a CDN, the embeddable option and serve
either icon-font or SVG from our server.

This change adds dependency on google/material icon repository and
create a build rule similar to that of the internal rule to create a SVG
bundle out of many SVG icons.