uploader: add `ServerInfo` protos and logic
Summary:
This commit adds an RPC definition by which the uploader can connect to
the frontend web server at the start of an upload session. This resolves
a number of outstanding issues:

  - The frontend can tell the uploader which backend server to connect
    to, rather than requiring a hard-coded endpoint in the uploader.
  - The frontend can tell the uploader how to generate experiment URLs,
    rather than requiring the backend server to provide this information
    (which it can’t, really, in general).
  - The frontend can check whether the uploader client is recent enough
    and instruct the end user to update if it’s not.
  - The frontend can warn the user about transient issues in case the
    service is down, degraded, under maintenance, etc.

An endpoint `https://tensorboard.dev/api/uploader` on the server will
provide this information.

Test Plan:
Unit tests suffice.

wchargin-branch: uploader-serverinfo-protos