analytics: send pathname on pageview event
Notice: this change only applies when Google Analytics is enabled which,
in OSS, it is not.

This change allows path like /experiment/abc to be logged as
/experiment/abc/scalar.

Do note that, according to Google Analaytics API doc, `ga('set', 'page',
arg)` is a way of modifying page url often to sanitize PII from the
path. This impacts all subsequent events like pageview or others.