uploader: add metadata to experiment listing RPC
Summary:
This extends the `StreamExperiments` RPC such that the client can
specify a set of additional metadata fields that the server should
include, like “creation time” or “number of scalar points”.

The format is both forward- and backward-compatible. Servers are
expected to send responses with both `experiment_ids` and `experiments`
until we drop support for clients that do not support `experiments`, at
which point they need only send `experiments`.

Test Plan:
Unit test added to simulate the future behavior of servers.

wchargin-branch: streamexperiments-metadata