core: security validator middleware
http_util enforces good practice in term of security to prevent content
sniffing and the content security policy. However, this is not
enforceable for other plugins.

With this middleware, we would like to tell developer to enhance the
security to protect users. When we go 3.0, we may convert the warns into
hard failures.