build: deprecate shasum in Vulcanization
As part of our security hardening and CSP support, we decided to
introduce script checksum as part of the Vulcanization that gets used at
the serving time using http_util.py. We understand that this imposes
a bit too stringent requirements for all and we decided to relax the CSP
comformance a little by trusting JavaScript from our server
(effectively, default-src 'self').

Note that we do not yet remove shasum related code from http_util
because it is being used by the Plugin HTML endpoint in application.py.
For that, we may want to use nonce or separate JavaScript file approach.