docs: remove outdated `bazel-bin` instructions
Summary:
Running Python binaries directly through `bazel-bin/` skips the Bazel
wrapper that sets up the environment correctly. This is not guaranteed
to work; users should run `bazel run //tensorboard --` instead.

We included the `bazel-bin/` form before `bazel run` was changed to
release the Bazel server lock, but this was fixed in 2017:
<https://github.com/bazelbuild/bazel/issues/2337>
…so there’s no longer really any reason to promote this form at all.

wchargin-branch: docs-no-bazel-bin