ng-tensorboard: Fix iframe based plugin rendering
Instead of JavaScript inside an iframe document, for security, we now
serve the document from the server.

Note that it was non-trivial to render the iframe in the Angular
template since it does not have virtual DOM and any prop changes would
trigger iframe's `src` to "change" which caused a flicker on the iframe.