black: manually format CRC table
Summary:
Per suggestion of @nfelt on #2967, we abstain from reformatting the CRC
table to one-per-line, for brevity and because this table is unlikely to
experience churn. We manually upcase the constants for consistency with
Black style.

Test Plan:
Running `black` on this file no longer changes the CRC table.

wchargin-branch: fmtoff-crc-table