pylint: remove unnecessary long-line suppressions
Summary:
In ec7e68d543a6, we removed `line-too-long` suppressions on import
lines. Most of the remaining `line-too-long` suppressions are also
unneeded, either because they’re around blocks of imports or because
none of the lines that they encapsulate is too long.

Test Plan:
The two remaining matches for `git grep line-too-long` are legitimate.

wchargin-branch: unsuppress-long-import-blocks