ng: material icon svg (#3017)
Unlike the default material-icon implementation, TensorBoard, which
tries to optimize to not make remote requests, bundles the SVG icons
with the pip package. To do so, we download the SVG from GitHub then use
a build rule to bundle the ones we use into our app. This change creates
the infrastructure and make use of the cog icon.

This is a manual revert of #2855.