[JS] upgrade dependencies (#3082)
In order to receive necessary security and bug fixes, we should seldom
upgrade our dependencies. This updates our shrinkwrapped dependencies
using `yarn upgrade`. Also updated our Angular dependencies on
package.json to be thorough.

About angular-metadata change:
`@angular/cli` introduced a schema that is not compatible with the ngc
in a patch release. As a result, we need to black list it from the
postinstall metadata/factory building. The "official" Bazel + Angular
repo already migrated to new AOT compiler using ivy and is not
representative of what we should do (it pinned to alpha release of
Angular 9.0).