ci: move linting to GitHub actions (#3102)
Summary:
GitHub Actions is a new first-party CI service offered by GitHub. It
requires no extra permissions. Its concurrency limits are appealing,
at 20 workflows per repo (1 workflow ≈ 1 commit) and concurrent jobs
ranging from 20 (free tier) to 180 (enterprise tier), with the option to
run on your own servers if this isn’t enough.

This commit adds a workflow definition to replace only the lint phase of
our Travis workflow, to take advantage of the improved latency: order of
seconds from pushing a commit to seeing actionable failure messages.
A full workflow for building and running tests is also possible, but is
too slow for comfort because it runs entirely without cache, pending
some proposed improvements to the caching support for GitHub Actions.

Test Plan:
Note that this commit triggers a GitHub Actions workflow that succeeds,
and that all Travis builds are a bit faster due to not running lint.

wchargin-branch: gh-actions-lint-only