backend: properly lock the locks
Summary:
The `plugin_event_accumulator` tries to maintain fine-grained locks for
each plugin, but didn’t properly lock its table of locks, and didn’t
lock around accesses to the table of plugins. This commit introduces a
coarse-grained lock to fix those problems.

Because all `dict`s involved are keyed only on Python strings, for which
`__hash__`, `__eq__`, and `__del__` are implemented in C, this should be
a no-op in CPython implementations. But correct concurrency is
[explicitly part of Google style][style], and without this patch some
subsequent changes would need to make this problem worse, so we might as
well just fix it Once and For All.

[style]: https://google.github.io/styleguide/pyguide.html#218-threading

Test Plan:
The primary safety concern due to this change is potential for liveness
failure. In theory, we should be safe because there’s a partial order
over locks: the new metalock is always acquired first. Empirically,
I ran this build of TensorBoard against a logdir with 4 GB of data and
8 concurrent reload threads plus an auto-reloading browser session, and
didn’t detect any issues.

wchargin-branch: mux-lock-locks