backend: fix locking for `_plugin_to_tag_to_content` (#3122)
Summary:
The `plugin_event_accumulator` tries to maintain fine-grained locks for
each plugin, but didn’t properly lock its table of locks, and didn’t
lock around accesses to the table of plugins. This commit replaces the
fine-grained locking with a simple coarse-grained lock.

Because all `dict`s involved are keyed only on Python strings, for which
`__hash__`, `__eq__`, and `__del__` are implemented in C, this should be
a no-op in CPython implementations. But correct concurrency is
[explicitly part of Google style][style], and without this patch some
subsequent changes would need to make this problem worse, so we might as
well just fix it Once and For All.

[style]: https://google.github.io/styleguide/pyguide.html#218-threading

Test Plan:
The primary safety concern due to this change is potential for liveness
failure, but this change basically contracts a number of locks into one,
so any lock waiting cycle that exists now would also have existed
previously.

wchargin-branch: mux-lock-locks