data: add `list_plugins` for `is_active` support (#3124)
Summary:
The `is_active` method of TensorBoard plugins hasn’t worked well with
data providers so far, because it doesn’t take an `experiment_id`
argument, and adding one would break backward compatibility. Instead, we
can pull `is_active` handling into the core application logic, where
both the data provider and experiment ID are still available.

A new data provider method `list_plugins` takes an experiment ID and
returns names of plugins with relevant data. A new (experimental) plugin
API lets plugins specify whose data they want to read, defaulting to
just their own. The `/data/plugins_listing` logic now joins the list of
plugins with data (from the provider) with each plugin’s list of data
that it cares about to determine which plugins are active. As a fallback
for plugins whose data is not yet understood by the data provider, we
also check `is_active` for plugins that don’t pass the first test.

The new data provider method is optional (for now). If not implemented,
this patch has no effect.

As a side benefit, this can speed up the plugins listing route by a fair
amount: we only make one targeted query against the multiplexer rather
than querying all its `Tensors` once per plugin, and we don’t need to
invoke `is_active` at all for plugins that are known to be active by
this first check.

Test Plan:
Change the `is_active` methods for the scalars, graphs, and histograms
plugins to return `False` rather than `True` when a data provider is
set. Then, run TensorBoard in generic data mode pointing at (a) the
scalars demo, (b) the histograms demo, and (c) the images demo. Note
that in case (a) only the scalars and graphs dashboards are active, in
case (b) only the histograms and distributions dashboards are active
(this demo does not write graphs), and in case (c) only the images and
graphs dashboards are active. Note that this behavior is the same across
generic and non-generic data modes. Note that the behavior prior to this
patch is restored if the implementation of `list_plugins` in the
multiplexer-backed data provider is removed, so this is a no-op until
data providers choose to implement this method.

wchargin-branch: data-list-active-plugins