ci: inspect cache size
Summary:
Our Travis caches are steadily growing. I suspect that the Pip cache of
TensorFlow nightlies plays a role here, as there are new packages every
day and the old ones are never evicted. Let’s find out.

This is a test PR just to trigger CI. (It would be possible to actually
merge some logging like this, but we’d want to make sure that it doesn’t
slow down the actual success reporting. Therefore:
DO NOT SUBMIT

wchargin-branch: ci-inspect-cache