Surface HTML import interop for plugin_lib
plugin_lib was initially designed to be used with npm package and a
bundler. However, because we have a need to use it under traditional
Polymer code that uses HTML import, we need to devise a way for the
lib to work with tf_web_library.

Changed our test to use the new interop layer.