scalars: promote generic data to stable
Summary:
At this point, the generic data APIs are a fully featured replacement
for the multiplexer APIs, including `is_active` support. We now enable
them by default. They can still be disabled with `--generic_data false`.

Test Plan:
Patch the `MultiplexerDataProvider` to return hard-coded descriptions so
that the generic and non-generic flows are distinguishable, then note
that `tensorboard` uses the generic data APIs with `--generic_data`
either set to `true` or unspecified, and not when set to `false`.

wchargin-branch: data-stabilize-scalars