text: add data provider support
Summary:
This patch teaches the text plugin about the generic data APIs using
text tensors. This has the implication that text tensors are required to
be not too large in size, which should be fine: the complete works of
Shakespeare fit in 5.5 MB, and text summaries should be expected to be
far smaller than that.

Test Plan:
Change `TensorBoardWSGIApp` to set `multiplexer=None` on the context,
and note that `--generic_data true` still has a working text plugin,
while `--generic_data false` is completely broken, so everything is
wired up through the new APIs.

wchargin-branch: generic-text