profile: replace static plugin with dynamic plugin (#3258)
Summary:
A working version of the dynamic profile plugin is live on stable PyPI!
It’s already a notable improvement over the static profile plugin,
because it includes a web components polyfill so that the trace viewer
actually works. Thus, we can now safely retire the legacy static plugin
and the trace viewer component (only used by the profile plugin).

Test Plan:
Launch TensorBoard in a fresh virtualenv with only `tf-nightly`,
pointing to a directory with profile data. Note that the “Profile”
plugin shows installation instructions rather than the old static
plugin. Copy and paste the installation instructions into a shell and
relaunch TensorBoard. Note that the dynamic plugin dashboard is now
available and active by default, and the installation instructions have
disappeared.

wchargin-branch: profile-dynamic-only