ci: test against Python 3.5 rather than 3.6
Summary:
We used to test on Python 2.7 and 3.6, which was sufficient to catch
most version-specific bugs in our support range. When we dropped support
for Python 2.7, we left CI at 3.6 only, but this means that we don’t
catch Python 3.5–specific failures. Running at Python 3.5 only should be
sufficient to mostly cover 3.5 through 3.7, especially given that the
TensorBoard team mostly uses Python 3.7 for daily development.

Test Plan:
Run in CI. The initial version of this commit is expected to fail due to
a bug in Python 3.5: <https://bugs.python.org/issue9232>.

wchargin-branch: ci-py35