Use cPython 3.5–compatible function definitions (#3320)
Summary:
In Python as of PEP 3102, function definitions may include keyword-only
arguments after a splat (`def foo(x, *, y)`). But cPython 3.5 has a
parser bug (<https://bugs.python.org/issue9232>) when trailing commas
are used with this syntax, as in `def foo(x, *, y,)`. This is fixed in
Python 3.6 and up, but not backported to 3.5, so for now we should make
sure to omit these trailing commas.

The change to `pyproject.toml` prevents Black from re-introducing these
trailing commas. We leave `py27` in the `pyproject.toml` to avoid a mass
reformat removing `u`-prefixes from Unicode string literals (which is
fine, but should be a separate change for Bazel `srcs_version`).

We don’t run our full test suite on Python 3.5 yet, due to unrelated
issues (primarily, lack of `assert_not_called` and that `json.loads`
does not accept `bytes`). For now, it suffices to run `flake8` on both
Python 3.5 and Python 3.7, because that at least parses all the code.

Test Plan:
That CI passes suffices.

wchargin-branch: py35-fundefs