ng: reconfigure metareducer to use injectable
Previously, even in the debug mode, the logger metaReducer had no
effect. When debugging the issue, it was found that the ngfactory which
is part of Angular's DI system prebuilds injectable for the META_REDUCER
and inlines an empty object [1].

[1]: To reproduce, please build without this commit and inspect
bazel-bin/tensorboard/webapp/app_module.ngfactory.js. Look for, in very
dense code, USER_PROVIDED_META_REDUCERS and notice that it provides
an empty array.