Buildifier format js.bzl for future edits (#3332)
With autoformatter for JS and Python, we got used to auto-formatting
everything including build file using Buildifier. Currently, due to the
formatting on the js.bzl, a single line change can make the overzealous
autoformatter change more lines than intended which is quite bad for
code review. This change simply autformats the js.bzl with buildifier
for future edits.