ng: reconfigure metareducer to use injectable (#3328)
Previously, even in the debug mode, the logger metaReducer had no
effect. When debugging the issue, it was found that the ngfactory which
is part of Angular's DI system prebuilds injectable for the META_REDUCER
and inlines an empty object [1]. Do note that this problem only manifests
when we try to configure metaReducer in a conditional. Also do note that
calling `isDevMode()` method is disallowed in AOT compiler, which is the
default compiler in Angular 9.

[1]: To reproduce, please build without this commit and inspect
bazel-bin/tensorboard/webapp/app_module.ngfactory.js. Look for, in very
dense code, USER_PROVIDED_META_REDUCERS and notice that it provides
an empty array.