Experimental plugins specified using ExperimentalPluginLoader (#3344)
Allow an administrator of a Tensorboard to mark certain plugins as "experimental" and hidden to users by default. Users would then need to add query parameter "experimentalPlugin=" to the URL to see the experimental plugin.

Introduce marker class ExperimentalPlugin which acts as an annotation on either TBPlugin or TBLoader that indicates a plugin is experimental. TensorBoardWSGIApp() uses isinstance() to determine if a plugin is experimental and passes the information as a list of names, "experimental_plugins", in its call to TensorBoardWSGI()

TensorBoardWSGI._serve_plugins_listing() filters out plugins that are included in experimental_plugins but are not specified in an experimentalPlugin query parameter.

There is logic in createRouter() that forwards "experimentalPlugin" query parameter values from the original TensorBoard request to the call to the plugins_listing endpoint.

Usage examples:

class ExperimentalImagePlugin(
  images_plugin.ImagesPlugin,
  experimental_plugin.ExperimentalPlugin
):
  pass

class ExperimentalDebuggerPluginLoader(
  debugger_plugin_loader.DebuggerPluginLoader,
  experimental_plugin.ExperimentalPlugin
):
  pass